Tpetra::CrsMatrix: local_matrix_type now derives SizeType from graph.
Tpetra::CrsMatrix's local_matrix_type typedef, a Kokkos::CrsMatrix
specialization, now derives its SizeType template parameter from
Tpetra::CrsGraph::local_graph_type (a Kokkos::StaticCrsGraph
specialization).  This means that if we ever want to change the value
of this template parameter, we only have to do so in one place
(Tpetra::CrsGraph).

Build/Test Cases Summary
Enabled Packages: Tpetra
Disabled Packages: FEI,Moertel,SEACAS,STK,PyTrilinos,CTrilinos,Optika,GlobiPack,OptiPack,RBGen,Rythmos,MOOCHO,Sundance,Didasko,FEApp
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_RELEASE_NO_CXX11 => FAILED: passed=845,notpassed=4 => Not ready to push! (66.35 min)
3) MPI_RELEASE_CLASSIC_SERIAL => FAILED: build failed => Not ready to push! (5.37 min)
Other local commits for this build/test group: 98b4996, a3ea0dc, 2d54342
WARNING: Forced the push!

+++ packages/tpetra/core/src/kokkos_refactor/Tpetra_KokkosRefactor_CrsGraph_def.hpp
@@ -577,7 +577,8 @@ namespace Tpetra {
           }
         }
         nodeMaxNumRowEntries_ =
-          std::max (d_ptrs(rlcid + 1) - d_ptrs(rlcid), nodeMaxNumRowEntries_);
+          std::max (static_cast<size_t> (d_ptrs(rlcid + 1) - d_ptrs(rlcid)),
+                    nodeMaxNumRowEntries_);
       }
     }
